refactor: organize COBOL compiler
