+++ compile/cobol/control.go
@@ -1,0 +1,29 @@
+package cobolcode
+
+import "fmt"
+
+import "mochi/ast"
+
+// compileIf emits a simple IF/ELSE chain.
+func (c *Compiler) compileIf(n *ast.Node) {
+	cond := c.expr(n.Children[0])
+	c.writeln(fmt.Sprintf("    IF %s", cond))
+	c.indent++
+	for _, st := range n.Children[1].Children {
+		c.compileNode(st)
+	}
+	c.indent--
+	if len(n.Children) > 2 {
+		c.writeln("    ELSE")
+		c.indent++
+		if n.Children[2].Kind == "if" {
+			c.compileIf(n.Children[2])
+		} else {
+			for _, st := range n.Children[2].Children {
+				c.compileNode(st)
+			}
+		}
+		c.indent--
+	}
+	c.writeln("    END-IF")
+}
